Conflicts:
	libs/kotext/styles/KoParagraphStyle.h

+++ libs/kotext/KoTextEditor.cpp
@@ -93,10 +93,11 @@ static bool isRightToLeft(const QString &text)
 /*Private*/
 
 KoTextEditor::Private::Private(KoTextEditor *qq, QTextDocument *document)
-    : q(qq),
-    document (document),
-    headCommand(0),
-    isBidiDocument(false)
+    : q(qq)
+    , document (document)
+    , headCommand(0)
+    , isBidiDocument(false)
+    , editProtectionCached(false)
 {
     caret = QTextCursor(document);
     editorState = NoOp;
@@ -471,6 +472,10 @@ void KoTextEditor::registerTrackedChange(QTextCursor &selection, KoGenChange::Ty
 
 void KoTextEditor::bold(bool bold)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     d->updateState(KoTextEditor::Private::Format, i18n("Bold"));
     QTextCharFormat format;
     format.setFontWeight(bold ? QFont::Bold : QFont::Normal);
@@ -483,6 +488,10 @@ void KoTextEditor::bold(bool bold)
 
 void KoTextEditor::italic(bool italic)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     d->updateState(KoTextEditor::Private::Format, i18n("Italic"));
     QTextCharFormat format;
     format.setFontItalic(italic);
@@ -495,6 +504,10 @@ void KoTextEditor::italic(bool italic)
 
 void KoTextEditor::underline(bool underline)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     d->updateState(KoTextEditor::Private::Format, i18n("Underline"));
     QTextCharFormat format;
     if (underline) {
@@ -513,6 +526,10 @@ void KoTextEditor::underline(bool underline)
 
 void KoTextEditor::strikeOut(bool strikeout)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     d->updateState(KoTextEditor::Private::Format, i18n("Strike Out"));
     QTextCharFormat format;
     if (strikeout) {
@@ -530,6 +547,10 @@ void KoTextEditor::strikeOut(bool strikeout)
 
 void KoTextEditor::setHorizontalTextAlignment(Qt::Alignment align)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     class Aligner : public BlockFormatVisitor
     {
     public:
@@ -548,6 +569,10 @@ void KoTextEditor::setHorizontalTextAlignment(Qt::Alignment align)
 
 void KoTextEditor::setVerticalTextAlignment(Qt::Alignment align)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     QTextCharFormat::VerticalAlignment charAlign = QTextCharFormat::AlignNormal;
     if (align == Qt::AlignTop)
         charAlign = QTextCharFormat::AlignSuperScript;
@@ -565,6 +590,10 @@ void KoTextEditor::setVerticalTextAlignment(Qt::Alignment align)
 
 void KoTextEditor::decreaseIndent()
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     class Indenter : public BlockFormatVisitor
     {
     public:
@@ -583,6 +612,10 @@ void KoTextEditor::decreaseIndent()
 
 void KoTextEditor::increaseIndent()
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     class Indenter : public BlockFormatVisitor
     {
     public:
@@ -625,6 +658,10 @@ public:
 
 void KoTextEditor::decreaseFontSize()
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     d->updateState(KoTextEditor::Private::Format, i18n("Decrease font size"));
     FontResizer sizer(FontResizer::Shrink);
     CharFormatVisitor::visitSelection(this, sizer, i18n("Decrease font size"));
@@ -633,6 +670,10 @@ void KoTextEditor::decreaseFontSize()
 
 void KoTextEditor::increaseFontSize()
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     d->updateState(KoTextEditor::Private::Format, i18n("Increase font size"));
     FontResizer sizer(FontResizer::Grow);
     CharFormatVisitor::visitSelection(this, sizer, i18n("Increase font size"));
@@ -641,6 +682,10 @@ void KoTextEditor::increaseFontSize()
 
 void KoTextEditor::setFontFamily(const QString &font)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     d->updateState(KoTextEditor::Private::Format, i18n("Set Font"));
     QTextCharFormat format;
     format.setFontFamily(font);
@@ -652,6 +697,10 @@ void KoTextEditor::setFontFamily(const QString &font)
 
 void KoTextEditor::setFontSize(qreal size)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     d->updateState(KoTextEditor::Private::Format, i18n("Set Font Size"));
     QTextCharFormat format;
     format.setFontPointSize(size);
@@ -663,6 +712,10 @@ void KoTextEditor::setFontSize(qreal size)
 
 void KoTextEditor::setTextBackgroundColor(const QColor &color)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     d->updateState(KoTextEditor::Private::Format, i18n("Set Background Color"));
     QTextCharFormat format;
     format.setBackground(QBrush(color));
@@ -674,6 +727,10 @@ void KoTextEditor::setTextBackgroundColor(const QColor &color)
 
 void KoTextEditor::setTextColor(const QColor &color)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     d->updateState(KoTextEditor::Private::Format, i18n("Set Text Color"));
     QTextCharFormat format;
     format.setForeground(QBrush(color));
@@ -685,6 +742,10 @@ void KoTextEditor::setTextColor(const QColor &color)
 
 void KoTextEditor::setStyle(KoCharacterStyle *style)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     Q_ASSERT(style);
     d->updateState(KoTextEditor::Private::Format, i18n("Set Character Style"));
     QTextCharFormat format;
@@ -697,6 +758,10 @@ void KoTextEditor::setStyle(KoCharacterStyle *style)
 
 void KoTextEditor::setStyle(KoParagraphStyle *style)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     d->updateState(KoTextEditor::Private::Format, i18n("Set Paragraph Style"));
     const int start = qMin(position(), anchor());
     const int end = qMax(position(), anchor());
@@ -731,6 +796,10 @@ void KoTextEditor::setDefaultFormat()
 
 void KoTextEditor::addBookmark(const QString &name)
 {//TODO changeTracking
+    if (isEditProtected()) {
+        return;
+    }
+
     d->updateState(KoTextEditor::Private::Custom, i18n("Insert Bookmark"));
     KoBookmark *bookmark = new KoBookmark(d->document);
     bookmark->setName(name);
@@ -770,6 +839,10 @@ void KoTextEditor::addBookmark(const QString &name)
 
 KoInlineObject *KoTextEditor::insertIndexMarker()
 {//TODO changeTracking
+    if (isEditProtected()) {
+        return 0;
+    }
+
     QTextBlock block = d->caret.block();
     if (d->caret.position() >= block.position() + block.length() - 1)
         return 0; // can't insert one at end of text
@@ -785,6 +858,10 @@ KoInlineObject *KoTextEditor::insertIndexMarker()
 
 void KoTextEditor::insertInlineObject(KoInlineObject *inliner)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     d->updateState(KoTextEditor::Private::Custom, i18n("Insert Variable"));
 
     int startPosition = d->caret.position();
@@ -808,6 +885,10 @@ void KoTextEditor::insertInlineObject(KoInlineObject *inliner)
 
 void KoTextEditor::insertFrameBreak()
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     d->updateState(KoTextEditor::Private::KeyPress, i18n("Insert Break"));
     QTextBlock block = d->caret.block();
     if (d->caret.position() == block.position() && block.length() > 0) { // start of parag
@@ -829,6 +910,10 @@ void KoTextEditor::insertFrameBreak()
 
 bool KoTextEditor::deleteInlineObjects(bool backward)
 {
+    if (isEditProtected()) {
+        return false;
+    }
+
     return d->deleteInlineObjects(backward);
 }
 
@@ -894,6 +979,10 @@ int KoTextEditor::columnNumber() const
 
 void KoTextEditor::deleteChar()
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     if (!d->caret.hasSelection() && d->caret.atEnd())
         return;
     if (!d->deleteInlineObjects(false) || d->caret.hasSelection()) {
@@ -908,6 +997,10 @@ void KoTextEditor::deleteChar()
 
 void KoTextEditor::deletePreviousChar()
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     if (!d->caret.hasSelection() && d->caret.atStart())
         return;
     if (!d->deleteInlineObjects(false) || d->caret.hasSelection()) {
@@ -935,19 +1028,127 @@ bool KoTextEditor::hasSelection()
     return d->caret.hasSelection();
 }
 
+// To figure out if a selection is write protected we need to traverse the entire document
+// as sections build up the protectiveness recursively.
+bool KoTextEditor::recursiveProtectionCheck(QTextFrame::iterator it)
+{
+    do {
+        QTextBlock block = it.currentBlock();
+        QTextTable *table = qobject_cast<QTextTable*>(it.currentFrame());
+        QTextFrame *subFrame = it.currentFrame();
+        if (table) {
+            // There are 4 ways this table can be selected:
+            //  - "before to mid"
+            //  - "mid to after"
+            //  - "complex mid to mid"
+            //  - "simple mid to mid"
+            // The 3 first are entire cells, the fourth is within a cell
+
+            if (d->caret.selectionStart() <= table->lastPosition()
+                && d->caret.selectionEnd() >= table->firstPosition()) {
+                // We have a selection somewhere 
+                QTextTableCell cell1 = table->cellAt(d->caret.selectionStart());
+                QTextTableCell cell2 = table->cellAt(d->caret.selectionEnd());
+                if (cell1 != cell2) {
+                    // And the selection is complex or entire table
+                    int selectionRow;
+                    int selectionColumn;
+                    int selectionRowSpan;
+                    int selectionColumnSpan;
+                    if (!cell1.isValid() || !cell2.isValid()) {
+                        // entire table
+                        selectionRow = selectionColumn = 0;
+                        selectionRowSpan = table->rows();
+                        selectionColumnSpan = table->columns();
+                    } else {
+                        d->caret.selectedTableCells(&selectionRow, &selectionRowSpan, &selectionColumn, &selectionColumnSpan);
+                    }
+
+                    for (int r = selectionRow; r < selectionRow + selectionRowSpan; r++) {
+                        for (int c = selectionColumn; c < selectionColumn + 
+                                    selectionColumnSpan; c++) {
+                            QTextTableCell cell = table->cellAt(r,c);
+                            if (cell.format().boolProperty(KoTableCellStyle::CellIsProtected)) {
+                                return true;
+                            }
+
+                            if (recursiveProtectionCheck(cell.begin())) {
+                                return true;
+                            }
+                        }
+                    }
+                } else {
+                    // And the selection is simple
+                    if (cell1.format().boolProperty(KoTableCellStyle::CellIsProtected)) {
+                        return true;
+                    }
+                    return recursiveProtectionCheck(cell1.begin());
+                }
+            }
+            if (d->caret.selectionEnd() <= table->lastPosition()) {
+                return false;
+            }
+        } if (subFrame) {
+        } else {
+            // TODO build up the section stack 
+
+            if (d->caret.selectionStart() < block.position() + block.length()
+                && d->caret.selectionEnd() >= block.position()) {
+                // We have a selection somewhere 
+                // TODO return true if block is protected by section
+            }
+
+            // TODO tear down the section stack 
+
+            if (d->caret.selectionEnd() < block.position() + block.length()) {
+                return false;
+            }
+        }
+        if (!it.atEnd()) {
+            ++it;
+        }
+    } while (!it.atEnd());
+    return false;
+}
+
+bool KoTextEditor::isEditProtected(bool useCached)
+{
+    if (useCached) {
+        if (! d->editProtectionCached) {
+            d->editProtected = recursiveProtectionCheck(d->document->rootFrame()->begin());
+            d->editProtectionCached = true;
+        }
+        return d->editProtected;
+    }
+    d->editProtectionCached = false;
+    return recursiveProtectionCheck(d->document->rootFrame()->begin());
+}
+
 void KoTextEditor::insertBlock()
 {
+    if (isEditProtected()) {
+        return;
+    }
+
 //TODO
 }
 
 void KoTextEditor::insertBlock(const QTextBlockFormat &format)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     Q_UNUSED(format)
 //TODO
 }
 
 void KoTextEditor::insertBlock(const QTextBlockFormat &format, const QTextCharFormat &charFormat)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     Q_UNUSED(format)
     Q_UNUSED(charFormat)
 //TODO
@@ -955,12 +1156,20 @@ void KoTextEditor::insertBlock(const QTextBlockFormat &format, const QTextCharFo
 
 void KoTextEditor::insertFragment(const QTextDocumentFragment &fragment)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     Q_UNUSED(fragment)
 //TODO
 }
 
 void KoTextEditor::insertTable(int rows, int columns)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     d->updateState(KoTextEditor::Private::Custom, i18n("Insert Table"));
     QTextTableFormat tableFormat;
 
@@ -1012,6 +1221,10 @@ void KoTextEditor::insertTable(int rows, int columns)
 
 void KoTextEditor::insertTableRowAbove()
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     QTextTable *table = d->caret.currentTable();
     if (table) {
         int changeId = 0;
@@ -1026,6 +1239,10 @@ void KoTextEditor::insertTableRowAbove()
 
 void KoTextEditor::insertTableRowBelow()
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     QTextTable *table = d->caret.currentTable();
     if (table) {
         int changeId = 0;
@@ -1040,6 +1257,10 @@ void KoTextEditor::insertTableRowBelow()
 
 void KoTextEditor::insertTableColumnLeft()
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     QTextTable *table = d->caret.currentTable();
     if (table) {
         int changeId = 0;
@@ -1054,6 +1275,10 @@ void KoTextEditor::insertTableColumnLeft()
 
 void KoTextEditor::insertTableColumnRight()
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     QTextTable *table = d->caret.currentTable();
     if (table) {
         int changeId = 0;
@@ -1068,6 +1293,10 @@ void KoTextEditor::insertTableColumnRight()
 
 void KoTextEditor::deleteTableColumn()
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     QTextTable *table = d->caret.currentTable();
     if (table) {
         int changeId = 0;
@@ -1083,6 +1312,10 @@ void KoTextEditor::deleteTableColumn()
 
 void KoTextEditor::deleteTableRow()
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     QTextTable *table = d->caret.currentTable();
     if (table) {
         KoChangeTracker *changeTracker = KoTextDocument(d->document).changeTracker();
@@ -1097,6 +1330,10 @@ void KoTextEditor::deleteTableRow()
 
 void KoTextEditor::mergeTableCells()
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     d->updateState(KoTextEditor::Private::Custom, i18n("Merge Cells"));
 
     QTextTable *table = d->caret.currentTable();
@@ -1110,6 +1347,10 @@ void KoTextEditor::mergeTableCells()
 
 void KoTextEditor::splitTableCells()
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     d->updateState(KoTextEditor::Private::Custom, i18n("Split Cells"));
 
     QTextTable *table = d->caret.currentTable();
@@ -1124,6 +1365,10 @@ void KoTextEditor::splitTableCells()
 
 void KoTextEditor::insertTableOfContents()
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     d->updateState(KoTextEditor::Private::Custom, i18n("Insert Table Of Contents"));
 
     QTextBlockFormat tocFormat;
@@ -1214,6 +1459,10 @@ KoInlineCite *KoTextEditor::insertCitation()
 
 void KoTextEditor::insertText(const QString &text)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     d->updateState(KoTextEditor::Private::KeyPress, i18n("Key Press"));
 
     //first we make sure that we clear the inlineObject charProperty, if we have no selection
@@ -1249,6 +1498,10 @@ void KoTextEditor::insertText(const QString &text)
 
 void KoTextEditor::insertText(const QString &text, const QTextCharFormat &format)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     Q_UNUSED(text)
     Q_UNUSED(format)
 //TODO
@@ -1256,24 +1509,37 @@ void KoTextEditor::insertText(const QString &text, const QTextCharFormat &format
 
 void KoTextEditor::mergeBlockCharFormat(const QTextCharFormat &modifier)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     Q_UNUSED(modifier)
 //TODO
 }
 
 void KoTextEditor::mergeBlockFormat(const QTextBlockFormat &modifier)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     Q_UNUSED(modifier)
 //TODO
 }
 
 void KoTextEditor::mergeCharFormat(const QTextCharFormat &modifier)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     Q_UNUSED(modifier)
 //TODO
 }
 
 bool KoTextEditor::movePosition(QTextCursor::MoveOperation operation, QTextCursor::MoveMode mode, int n)
 {
+    d->editProtectionCached = false;
     bool b = d->caret.movePosition (operation, mode, n);
     emit cursorPositionChanged();
     return b;
@@ -1281,6 +1547,10 @@ bool KoTextEditor::movePosition(QTextCursor::MoveOperation operation, QTextCurso
 
 void KoTextEditor::newLine()
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     d->updateState(KoTextEditor::Private::Custom, i18n("Line Break"));
     if (d->caret.hasSelection())
         d->deleteInlineObjects();
@@ -1357,6 +1627,10 @@ int KoTextEditor::position() const
 
 void KoTextEditor::removeSelectedText()
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     d->caret.removeSelectedText();
     emit cursorPositionChanged();
 }
@@ -1389,30 +1663,47 @@ int KoTextEditor::selectionStart() const
 
 void KoTextEditor::setBlockCharFormat(const QTextCharFormat &format)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     Q_UNUSED(format)
 //TODO
 }
 
 void KoTextEditor::setBlockFormat(const QTextBlockFormat &format)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     Q_UNUSED(format)
 //TODO
 }
 
 void KoTextEditor::setCharFormat(const QTextCharFormat &format)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     Q_UNUSED(format)
 //TODO
 }
 
 void KoTextEditor::setTableFormat(const QTextTableFormat &format)
 {
+    if (isEditProtected()) {
+        return;
+    }
+
     Q_UNUSED(format)
 //TODO
 }
 
 void KoTextEditor::setPosition(int pos, QTextCursor::MoveMode m)
 {
+    d->editProtectionCached = false;
     d->caret.setPosition (pos, m);
     emit cursorPositionChanged();
 }
